fix(V-04): move DividendValuationRules from Application to Domain
RequiredYield and DividendYearsLookback are business rules (what yield
is acceptable, how many years to look back) — they encode policy that
belongs in the Domain, not in Application configuration.

Moved to Financial.Domain/Rules/DividendValuationRules.cs with namespace
Financial.Domain.Rules. DividendService and MainWindow.xaml.cs updated
to import from the new location. Financial.App.csproj gains an explicit
ProjectReference to Financial.Domain since it now consumes Domain types
directly.
