+++ Financial.Application/Services/DividendService.cs
@@ -1,7 +1,7 @@
-using Financial.Application.Configuration;
 using Financial.Application.DTOs;
 using Financial.Application.Interfaces;
 using Financial.Domain.Entities;
+using Financial.Domain.Rules;
 using System;
 using System.Collections.Generic;
 using System.Linq;
